+++ lib/ui/painting/gradient.h
@@ -32,7 +32,7 @@ class CanvasGradient : public Shader {
                   const tonic::Int32List& colors,
                   const tonic::Float32List& color_stops,
                   SkTileMode tile_mode,
-                  const tonic::Float64List& matrix4);
+                  const tonic::Float32List& matrix4);
 
   void initRadial(double center_x,
                   double center_y,
@@ -40,7 +40,7 @@ class CanvasGradient : public Shader {
                   const tonic::Int32List& colors,
                   const tonic::Float32List& color_stops,
                   SkTileMode tile_mode,
-                  const tonic::Float64List& matrix4);
+                  const tonic::Float32List& matrix4);
 
   void initSweep(double center_x,
                  double center_y,
@@ -49,7 +49,7 @@ class CanvasGradient : public Shader {
                  SkTileMode tile_mode,
                  double start_angle,
                  double end_angle,
-                 const tonic::Float64List& matrix4);
+                 const tonic::Float32List& matrix4);
 
   void initTwoPointConical(double start_x,
                            double start_y,
@@ -60,7 +60,7 @@ class CanvasGradient : public Shader {
                            const tonic::Int32List& colors,
                            const tonic::Float32List& color_stops,
                            SkTileMode tile_mode,
-                           const tonic::Float64List& matrix4);
+                           const tonic::Float32List& matrix4);
 
   static void RegisterNatives(tonic::DartLibraryNatives* natives);
