+++ lib/ui/semantics/semantics_update_builder.h
@@ -48,7 +48,7 @@ class SemanticsUpdateBuilder
                   std::string increasedValue,
                   std::string decreasedValue,
                   int textDirection,
-                  const tonic::Float64List& transform,
+                  const tonic::Float32List& transform,
                   const tonic::Int32List& childrenInTraversalOrder,
                   const tonic::Int32List& childrenInHitTestOrder,
                   const tonic::Int32List& customAccessibilityActions);
